refactor: move ConvertToModels from db to models package
