+++ db/redhat.go
@@ -3,13 +3,11 @@ package db
 import (
 	"errors"
 	"fmt"
-	"strings"
 	"time"
 
 	"github.com/inconshreveable/log15"
 	"github.com/spf13/viper"
 	"github.com/vulsio/gost/models"
-	"github.com/vulsio/gost/util"
 	"golang.org/x/xerrors"
 	pb "gopkg.in/cheggaaa/pb.v1"
 	"gorm.io/gorm"
@@ -149,12 +147,7 @@ func (r *RDBDriver) GetUnfixedCvesRedhat(major, pkgName string, ignoreWillNotFix
 }
 
 // InsertRedhat :
-func (r *RDBDriver) InsertRedhat(cveJSONs []models.RedhatCVEJSON) (err error) {
-	cves, err := ConvertRedhat(cveJSONs)
-	if err != nil {
-		return err
-	}
-
+func (r *RDBDriver) InsertRedhat(cves []models.RedhatCVE) (err error) {
 	if err := r.deleteAndInsertRedhat(cves); err != nil {
 		return xerrors.Errorf("Failed to insert RedHat CVE data. err: %w", err)
 	}
@@ -214,96 +207,4 @@ func (r *RDBDriver) deleteAndInsertRedhat(cves []models.RedhatCVE) (err error) {
 	bar.Finish()
 
 	return nil
-}
-
-// ConvertRedhat :
-func ConvertRedhat(cveJSONs []models.RedhatCVEJSON) (cves []models.RedhatCVE, err error) {
-	for _, cve := range cveJSONs {
-		details := []models.RedhatDetail{}
-		for _, d := range cve.Details {
-			d = util.TrimSpaceNewline(d)
-			details = append(details, models.RedhatDetail{Detail: d})
-		}
-
-		references := []models.RedhatReference{}
-		for _, r := range cve.References {
-			r = util.TrimSpaceNewline(r)
-			references = append(references, models.RedhatReference{Reference: r})
-		}
-
-		cve.Bugzilla.Description = util.TrimSpaceNewline(cve.Bugzilla.Description)
-		cve.Statement = util.TrimSpaceNewline(cve.Statement)
-
-		var publicDate time.Time
-		if cve.PublicDate != "" {
-			if strings.HasSuffix(cve.PublicDate, "Z") {
-				publicDate, err = time.Parse("2006-01-02T15:04:05Z", cve.PublicDate)
-			} else {
-				publicDate, err = time.Parse("2006-01-02T15:04:05", cve.PublicDate)
-			}
-			if err != nil {
-				return nil, xerrors.Errorf("Failed to parse date. date: %s err: %w", cve.PublicDate, err)
-			}
-		}
-
-		// TODO: more efficient
-		c := models.RedhatCVE{
-			ThreatSeverity:       cve.ThreatSeverity,
-			PublicDate:           publicDate,
-			Bugzilla:             cve.Bugzilla,
-			Cvss:                 cve.Cvss,
-			Cvss3:                cve.Cvss3,
-			Iava:                 cve.Iava,
-			Cwe:                  cve.Cwe,
-			Statement:            cve.Statement,
-			Acknowledgement:      cve.Acknowledgement,
-			Mitigation:           cve.Mitigation,
-			AffectedRelease:      cve.AffectedRelease,
-			PackageState:         cve.PackageState,
-			Name:                 cve.Name,
-			DocumentDistribution: cve.DocumentDistribution,
-
-			Details:    details,
-			References: references,
-		}
-		cves = append(cves, c)
-	}
-	return cves, nil
-}
-
-// ClearIDRedhat :
-func ClearIDRedhat(cve *models.RedhatCVE) {
-	cve.ID = 0
-	cve.Bugzilla.RedhatCVEID = 0
-	cve.Cvss.RedhatCVEID = 0
-	cve.Cvss3.RedhatCVEID = 0
-
-	affectedReleases := cve.AffectedRelease
-	cve.AffectedRelease = []models.RedhatAffectedRelease{}
-	for _, a := range affectedReleases {
-		a.RedhatCVEID = 0
-		cve.AffectedRelease = append(cve.AffectedRelease, a)
-	}
-
-	packageState := cve.PackageState
-	cve.PackageState = []models.RedhatPackageState{}
-	for _, p := range packageState {
-		p.RedhatCVEID = 0
-		cve.PackageState = append(cve.PackageState, p)
-	}
-
-	details := cve.Details
-	cve.Details = []models.RedhatDetail{}
-	for _, d := range details {
-		d.RedhatCVEID = 0
-		cve.Details = append(cve.Details, d)
-	}
-
-	references := cve.References
-	cve.References = []models.RedhatReference{}
-	for _, r := range references {
-		r.RedhatCVEID = 0
-		cve.References = append(cve.References, r)
-	}
-
 }
