+++ notifier/redhat.go
@@ -1,4 +1,4 @@
-package util
+package notifier
 
 import (
 	"fmt"
@@ -8,6 +8,43 @@ import (
 	"github.com/vulsio/gost/config"
 	"github.com/vulsio/gost/models"
 )
+
+// ClearIDRedhat :
+func ClearIDRedhat(cve *models.RedhatCVE) {
+	cve.ID = 0
+	cve.Bugzilla.RedhatCVEID = 0
+	cve.Cvss.RedhatCVEID = 0
+	cve.Cvss3.RedhatCVEID = 0
+
+	affectedReleases := cve.AffectedRelease
+	cve.AffectedRelease = []models.RedhatAffectedRelease{}
+	for _, a := range affectedReleases {
+		a.RedhatCVEID = 0
+		cve.AffectedRelease = append(cve.AffectedRelease, a)
+	}
+
+	packageState := cve.PackageState
+	cve.PackageState = []models.RedhatPackageState{}
+	for _, p := range packageState {
+		p.RedhatCVEID = 0
+		cve.PackageState = append(cve.PackageState, p)
+	}
+
+	details := cve.Details
+	cve.Details = []models.RedhatDetail{}
+	for _, d := range details {
+		d.RedhatCVEID = 0
+		cve.Details = append(cve.Details, d)
+	}
+
+	references := cve.References
+	cve.References = []models.RedhatReference{}
+	for _, r := range references {
+		r.RedhatCVEID = 0
+		cve.References = append(cve.References, r)
+	}
+
+}
 
 // DiffRedhat returns the difference between the old and new CVE information
 func DiffRedhat(old, new *models.RedhatCVE, config config.RedhatWatchCve) (body string) {
